Move TacoSignature out of Porter and into sign.ts since Porter no longer handles unencrypted responses. Export relevant types from sign module.
